finish alert dialog for deleting customers and plans

+++ src/app/customer-list/customer-list.component.ts
@@ -4,6 +4,8 @@ import { ActivatedRoute } from '@angular/router';
 import { CustomerService } from '../customer.service';
 import { Observable } from 'rxjs';
 import { getRandomString } from 'selenium-webdriver/safari';
+import { MatDialog, MatDialogConfig } from '@angular/material';
+import { MyDialogComponent } from '../my-dialog/my-dialog.component';
 
 @Component({
   selector: 'app-customer-list',
@@ -16,7 +18,8 @@ export class CustomerListComponent implements OnInit {
   paramsSubscription;
   selectedCustomer: Customer;
   constructor(private customerService: CustomerService,
-    private activatedRoute: ActivatedRoute) { }
+    private activatedRoute: ActivatedRoute,
+    public dialog: MatDialog) { }
 
 
 
@@ -42,14 +45,43 @@ export class CustomerListComponent implements OnInit {
 
   }
   
-  deleteCustomer(id: string) {
-    console.log("customer id: " + id);
-   // this.customerService.deleteCustomer(id).subscribe();
-    this.customerService.deleteCustomer(id).subscribe(data=>{
-      console.log("esdf");
-      this.getCustomers("all");
+//  deleteCustomer(id: string) {
+  //  this.openModal(id);
+  //   console.log("customer id: " + id);
+  //  // this.customerService.deleteCustomer(id).subscribe();
+  //   this.customerService.deleteCustomer(id).subscribe(data=>{
+  //     console.log("esdf");
+  //     this.getCustomers("all");
+  //   });
+ // }
+
+ deleteCustomer(id: string) {
+    const dialogConfig = new MatDialogConfig();
+    var customerName: string;
+
+    this.customerService.getCustomer(id).toPromise().then(customer => { 
+      this.selectedCustomer = customer;
+      console.log("--");
+      console.log(customer);
+      customerName = "Delete customer: " + customer['firstName'] + ' ' +customer['lastName'] ;
+    }).then(()=>{
+      dialogConfig.disableClose = true;
+      dialogConfig.autoFocus = true;
+      dialogConfig.data = {
+      id: 1,
+      title: customerName
+      };
+    }).then(()=>{
+      const dialogRef = this.dialog.open(MyDialogComponent, dialogConfig);
+      dialogRef.afterClosed().subscribe(result => {
+        if(result){
+           this.customerService.deleteCustomer(id).subscribe(data=>{
+           this.getCustomers("all");
+         });
+        }
+       console.log("Dialog was closed!")
+       console.log(result)
+      });
     });
-
   }
-
 }
